[tvOS] Fix smaller issues in code imported from C25 (#8013)
Fix a bunch of erros and warnings in the code that came from C25. There
are many small changes that would be too annoying to send in separate
PRs:

- Fix some unsigned-to-signed comparisons
- Fix some narrowing (float to int) issues
- Avoid shadowing variables in `game_controller_input_device.mm`
- Remove some trailing or wrong semicolons in declarations
- Remove unused constants and variables
- Remove a break after a return statement in a switch clause

Bug: 459507205
Bug: 456692793

+++ starboard/tvos/shared/media/playback_capabilities.h
@@ -33,7 +33,7 @@ class PlaybackCapabilities {
   // Returns true if the device is Apple TV 4K.
   static bool IsAppleTV4K();
   // Return audio output configuration.
-  static bool GetAudioConfiguration(int index,
+  static bool GetAudioConfiguration(size_t index,
                                     SbMediaAudioConfiguration* configuration);
   // Force to reload audio configurations.
   static void ReloadAudioConfigurations();
